Broke the chat module up into separate component files Added a missing constant that actions.js was looking for Fixed a typo in userList.js that was causing some issues Added a .gitignore
