+++ src/constants.js
@@ -15,4 +15,7 @@ module.exports = {
 	LOGGED_IN: "LOGGED_IN",
 	ANONYMOUS: "ANONYMOUS",
 	AWAITING_AUTH_RESPONSE: "AWAITING_AUTH_RESPONSE",
+
+	// ERROR
+	DISPLAY_ERROR: "DISPLAY_ERROR"
 };
